+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - CPX-AP: Added diagnosis example
 
+
+## v0.5.0 - 17.06.24
+### Added
+- Diagnosis for CpxAp and ApModule. This includes a ModuleInformation dataclass and builder as well as some new functions
+- CpxAp: Read the global diagnosis state from the cpx-ap system with `CpxAp.read_global_diagnosis_state()`
+- CpxAp: Read the count of active diagnosis from the cpx-ap system with `CpxAp.read_active_diagnosis_count()`
+- CpxAp: Read the module index with the latest diagnosis from the cpx-ap system with `CpxAp.read_latest_diagnosis_index()`
+- CpxAp: Read the latest diagnosis code from the cpx-ap system with `CpxAp.read_latest_diagnosis_code()`
+- CpxAp: Added INT16 and UINT16 support for MPA modules in read/write functions, expanded system tests
+- ApModule: Read the diagnosis code from the module with `ApModule.read_diagnosis_code()`
+- ApModule: Read the diagnosis information from the module with `ApModule.read_diagnosis_information()`
+
 ## v0.4.2 - 29.05.24
 ### Changed
 - Updated links to public repo